added buy currency functionality

+++ Server/Server/Integrations/NbpApi/Models/NbpCurrency.cs
@@ -4,7 +4,7 @@ public sealed class NbpCurrency
 {
     public required string Currency { get; init; }
     public required string Code { get; init; }
-    public double Mid { get; init; }
+    public decimal Mid { get; init; }
     public decimal Bid { get; init; }
     public decimal Ask { get; init; }  
 }
